requirements.txt
pymobiledevice3
python-dotenv
